EES-7570 add new prototype route for homepage redesign
