+++ src/explore-education-statistics-frontend/src/pages/prototype/homepage-redesign.tsx
@@ -1,0 +1,1 @@
+export { default } from '@frontend/modules/home/HomePageRedesign';
